+++ tests_and_analysis/test/script_tests/test_powder_map.py
@@ -59,6 +59,11 @@ powder_map_params_from_phonopy = [
      *quick_calc_params],
     [nacl_prim_fc_file, '--temperature=1000', '--weighting=coherent',
      *quick_calc_params]]
+powder_map_params_brille = [[graphite_fc_file, '--use-brille',
+                             '--brille-npts', '10',
+                             '--brille-grid-type', 'nest',
+                             '-w', 'coherent',
+                             *quick_calc_params]]
 
 class TestRegression:
 
@@ -73,7 +78,8 @@ class TestRegression:
         # Ensure figures are closed
         matplotlib.pyplot.close('all')
 
-    def run_powder_map_and_test_result(self, powder_map_args):
+    def run_powder_map_and_test_result(
+            self, powder_map_args, keys_to_omit=['x_ticklabels']):
         euphonic.cli.powder_map.main(powder_map_args)
 
         matplotlib.pyplot.gcf().tight_layout()  # Force tick labels to be set
@@ -83,8 +89,8 @@ class TestRegression:
             expected_image_data = json.load(expected_data_file)[
                 args_to_key(powder_map_args)]
         for key, expected_val in expected_image_data.items():
-            # We don't care about the details of tick labels for powder map
-            if key == 'x_ticklabels':
+            if key in keys_to_omit:
+                # We don't care about the details of tick labels for powder map
                 pass
             elif key == 'extent':
                 # Lower bound of y-data (energy) varies by up to ~2e-6 on
@@ -106,13 +112,66 @@ class TestRegression:
         self.run_powder_map_and_test_result(powder_map_args)
 
     @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
     @pytest.mark.parametrize(
         'powder_map_args', powder_map_params_from_phonopy)
     def test_powder_map_plot_image_from_phonopy(
             self, inject_mocks, powder_map_args):
         self.run_powder_map_and_test_result(powder_map_args)
 
+    @pytest.mark.brille
+    @pytest.mark.multiple_extras
+    @pytest.mark.parametrize(
+        'powder_map_args', powder_map_params_brille)
+    def test_powder_map_plot_image_with_brille(
+            self, inject_mocks, powder_map_args):
+        # Different numerical results on different platforms
+        # unless a very dense, computationally expensive grid
+        # is used. Just check that the program runs and a plot
+        # is produced, by omitting check of 'data_1', 'data_2'
+        self.run_powder_map_and_test_result(
+            powder_map_args,
+            keys_to_omit=['x_ticklabels', 'data_1', 'data_2'])
+
+    @pytest.mark.brille
+    @pytest.mark.multiple_extras
+    @pytest.mark.parametrize('powder_map_args, expected_kwargs', [
+        (['--use-brille', '--brille-npts', '25', '--disable-c'],
+         {'grid_npts': 25, 'grid_type': 'trellis',
+          'interpolation_kwargs': {'use_c': False}}),
+        (['--use-brille', '--brille-grid-type', 'mesh', '--use-c',
+          '--n-threads', '2'],
+          {'grid_type': 'mesh', 'interpolation_kwargs': {'use_c': True,
+                                                         'n_threads': 2}})
+    ])
+    def test_brille_interpolator_from_force_constants_kwargs_passed(
+            self, inject_mocks, mocker, powder_map_args, expected_kwargs):
+        from euphonic.brille import BrilleInterpolator
+        # Stop execution once from_fc has been called - we're only
+        # checking here that the correct arguments have been passed
+        # through
+        class MockException(Exception):
+            pass
+        mock = mocker.patch.object(BrilleInterpolator, 'from_force_constants',
+                                   side_effect=MockException())
+        try:
+            euphonic.cli.powder_map.main(
+                [graphite_fc_file] + powder_map_args + quick_calc_params)
+        except MockException:
+            pass
+        default_interp_kwargs =  {'asr': None, 'dipole_parameter': 1.0,
+                                  'n_threads': None, 'use_c': None}
+        default_kwargs = {'grid_type': 'trellis', 'grid_npts': 5000,
+                          'grid_density': None}
+        expected_interp_kwargs = {
+            **default_interp_kwargs,
+            **expected_kwargs.pop('interpolation_kwargs', {})}
+        expected_kwargs = {**default_kwargs, **expected_kwargs}
+        expected_kwargs['interpolation_kwargs'] = expected_interp_kwargs
+        assert mock.call_args[1] == expected_kwargs
+
     @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
     @pytest.mark.parametrize('powder_map_args', [
         [nacl_prim_fc_file, '--save-to'],
         [nacl_prim_fc_file, '-s']])
@@ -138,10 +197,11 @@ class TestRegression:
             euphonic.cli.powder_map.main(powder_map_args)
 
     @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
     @pytest.mark.parametrize('powder_map_args', [
-        [os.path.join(
-            get_data_path(), 'qpoint_phonon_modes',
-            'quartz', 'quartz_bandstructure_qpoint_phonon_modes.json')]])
+        [get_data_path(
+            'qpoint_phonon_modes', 'quartz',
+            'quartz_bandstructure_qpoint_phonon_modes.json')]])
     def test_qpoint_modes_raises_type_error(self, powder_map_args):
         with pytest.raises(TypeError):
             euphonic.cli.powder_map.main(powder_map_args)
@@ -155,7 +215,20 @@ class TestRegression:
         assert err.type == SystemExit
         assert err.value.code == 2
 
+    @pytest.mark.brille
     @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
+    @pytest.mark.parametrize('powder_map_args', [
+        [nacl_prim_fc_file, '--use-brille', '--brille-grid-type', 'grid']])
+    def test_invalid_brille_grid_raises_causes_exit(self, powder_map_args):
+        # Argparse should call sys.exit on invalid choices
+        with pytest.raises(SystemExit) as err:
+            euphonic.cli.powder_map.main(powder_map_args)
+        assert err.type == SystemExit
+        assert err.value.code == 2
+
+    @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
     @pytest.mark.parametrize('powder_map_args', [
         [nacl_prim_fc_file, '-w=coherent', '--pdos', 'Na']])
     def test_coherent_weighting_and_pdos_raises_value_error(
@@ -163,6 +236,8 @@ class TestRegression:
         with pytest.raises(ValueError):
             euphonic.cli.powder_map.main(powder_map_args + quick_calc_params)
 
+    @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
     @pytest.mark.parametrize('powder_map_args', [
         [nacl_prim_fc_file, '--pdos']])
     def test_no_pdos_args_raises_causes_exit(self, powder_map_args):
@@ -171,6 +246,8 @@ class TestRegression:
         assert err.type == SystemExit
         assert err.value.code == 2
 
+    @pytest.mark.phonopy_reader
+    @pytest.mark.multiple_extras
     @pytest.mark.parametrize('powder_map_args', [
         [nacl_prim_fc_file, '--pdos', 'Na', 'Cl']])
     def test_multiple_pdos_args_raises_causes_exit(self, powder_map_args):
@@ -192,7 +269,8 @@ def test_regenerate_powder_map_data(_):
         json_data = {}
 
     for powder_map_param in (powder_map_params
-                             + powder_map_params_from_phonopy):
+                             + powder_map_params_from_phonopy
+                             + powder_map_params_brille):
         # Generate current figure for us to retrieve with gcf
         euphonic.cli.powder_map.main(powder_map_param)